Use subcommand pattern for install-tool
