+++ src/dotnet/BuiltInCommandsCatalog.cs
@@ -20,6 +20,7 @@ using Microsoft.DotNet.Tools.Store;
 using Microsoft.DotNet.Tools.Test;
 using Microsoft.DotNet.Tools.VSTest;
 using System.Collections.Generic;
+using Microsoft.DotNet.Tools.Install;
 
 namespace Microsoft.DotNet.Cli
 {
