+++ src/dotnet/commands/dotnet-install/InstallCommandParser.cs
@@ -11,29 +11,9 @@ namespace Microsoft.DotNet.Cli
         {
             return Create.Command(
                 "install", "",
-                Accept.NoArguments(), CommonOptions.HelpOption(), InstallTool());
-        }
-
-        private static Command InstallTool()
-        {
-            return Create.Command("tool",
-                "Install tool",
-                Accept.ExactlyOneArgument(o => "packageId")
-                    .With(name: "packageId",
-                        description: "Package Id in NuGet"),
-                Create.Option(
-                    "--version",
-                    "Version of the package in NuGet",
-                    Accept.ExactlyOneArgument()),
-                Create.Option(
-                    "--configfile",
-                    "NuGet configuration file",
-                    Accept.ExactlyOneArgument()),
-                Create.Option(
-                    "-f|--framework",
-                    "Target Framework Moniker (TFM) for the tools to install",
-                    Accept.ExactlyOneArgument()),
-                CommonOptions.HelpOption());
+                Accept.NoArguments(),
+                CommonOptions.HelpOption(),
+                InstallToolCommandParser.InstallTool());
         }
     }
 }
